+++ designer/client/src/http/HttpService.ts
@@ -3,7 +3,7 @@ import { AxiosError, AxiosResponse } from "axios";
 import FileSaver from "file-saver";
 import i18next from "i18next";
 import { Moment } from "moment";
-import { ProcessingType, SettingsData, ValidationData, ValidationRequest } from "../actions/nk";
+import { Position, ProcessingType, SettingsData, ValidationData, ValidationRequest } from "../actions/nk";
 import { GenericValidationRequest, TestAdhocValidationRequest } from "../actions/nk/adhocTesting";
 import api from "../api";
 import { UserData } from "../common/models/User";
@@ -33,10 +33,12 @@ import { EventTrackingSelectorType, EventTrackingType } from "../containers/even
 import { BackendNotification } from "../containers/Notifications";
 import { ProcessCounts } from "../reducers/graph";
 import { AuthenticationSettings } from "../reducers/settings";
-import { Expression, NodeType, ProcessAdditionalFields, ProcessDefinitionData, ScenarioGraph, VariableTypes } from "../types";
+import { Expression, LayoutData, NodeType, ProcessAdditionalFields, ProcessDefinitionData, ScenarioGraph, VariableTypes } from "../types";
 import { Instant, WithId } from "../types/common";
 import { fixAggregateParameters, fixBranchParametersTemplate } from "./parametersUtils";
 import { handleAxiosError } from "../devHelpers";
+import { Dimensions, StickyNote } from "../common/StickyNote";
+import { STICKY_NOTE_DEFAULT_COLOR } from "../components/graph/EspNode/stickyNote";
 
 type HealthCheckProcessDeploymentType = {
     status: string;
@@ -124,9 +126,10 @@ export type ComponentUsageType = {
     lastAction: ProcessActionType;
 };
 
-type NotificationActions = {
+export type NotificationActions = {
     success(message: string): void;
     error(message: string, error: string, showErrorText: boolean): void;
+    warn(message: string): void;
 };
 
 export interface TestProcessResponse {
@@ -679,6 +682,57 @@ class HttpService {
                     true,
                 ),
             );
+        return promise;
+    }
+
+    addStickyNote(scenarioName: string, scenarioVersionId: number, position: Position, dimensions: Dimensions) {
+        const promise = api.post(`/processes/${encodeURIComponent(scenarioName)}/stickyNotes`, {
+            scenarioVersionId,
+            content: "",
+            layoutData: position,
+            color: STICKY_NOTE_DEFAULT_COLOR, //TODO add config for default sticky note color? For now this is default.
+            dimensions: dimensions,
+        });
+        promise.catch((error) => {
+            const errorMsg: string = error?.response?.data;
+            this.#addError("Failed to add sticky note" + (errorMsg ? ": " + errorMsg : ""), error, true);
+        });
+        return promise;
+    }
+
+    deleteStickyNote(scenarioName: string, stickyNoteId: number) {
+        const promise = api.delete(`/processes/${encodeURIComponent(scenarioName)}/stickyNotes/${stickyNoteId}`);
+        promise.catch((error) =>
+            this.#addError(
+                i18next.t("notification.error.failedToDeleteStickyNote", `Failed to delete sticky note with id: ${stickyNoteId}`),
+                error,
+                true,
+            ),
+        );
+        return promise;
+    }
+
+    updateStickyNote(scenarioName: string, scenarioVersionId: number, stickyNote: StickyNote) {
+        const promise = api.put(`/processes/${encodeURIComponent(scenarioName)}/stickyNotes`, {
+            noteId: stickyNote.noteId,
+            scenarioVersionId,
+            content: stickyNote.content,
+            layoutData: stickyNote.layoutData,
+            color: stickyNote.color,
+            dimensions: stickyNote.dimensions,
+        });
+        promise.catch((error) => {
+            const errorMsg = error?.response?.data;
+            this.#addError("Failed to update sticky note" + errorMsg ? ": " + errorMsg : "", error, true);
+        });
+        return promise;
+    }
+
+    getStickyNotes(scenarioName: string, scenarioVersionId: number): Promise<AxiosResponse<StickyNote[]>> {
+        const promise = api.get(`/processes/${encodeURIComponent(scenarioName)}/stickyNotes?scenarioVersionId=${scenarioVersionId}`);
+        promise.catch((error) =>
+            this.#addError(i18next.t("notification.error.failedToGetStickyNotes", "Failed to get sticky notes"), error, true),
+        );
         return promise;
     }