+++ designer/client/src/reducers/graph/reducer.ts
@@ -16,9 +16,12 @@ import { selectionState } from "./selectionState";
 import { GraphState } from "./types";
 import {
     addNodesWithLayout,
+    addStickyNotesWithLayout,
     adjustBranchParametersAfterDisconnect,
     createEdge,
     enrichNodeWithProcessDependentData,
+    prepareNewStickyNotesWithLayout,
+    removeStickyNoteFromLayout,
     updateAfterNodeDelete,
     updateLayoutAfterNodeIdChange,
 } from "./utils";
@@ -227,6 +230,18 @@ const graphReducer: Reducer<GraphState> = (state = emptyGraphState, action) => {
                 nodes: action.nodes,
                 layout: action.layout,
             });
+        }
+        case "STICKY_NOTES_UPDATED": {
+            const { stickyNotes, layout } = prepareNewStickyNotesWithLayout(state, action.stickyNotes);
+            return {
+                ...addStickyNotesWithLayout(state, { stickyNotes, layout }),
+            };
+        }
+        case "STICKY_NOTE_DELETED": {
+            const { stickyNotes, layout } = removeStickyNoteFromLayout(state, action.stickyNoteId);
+            return {
+                ...addStickyNotesWithLayout(state, { stickyNotes, layout }),
+            };
         }
         case "NODES_WITH_EDGES_ADDED": {
             const { nodes, layout, idMapping, processDefinitionData, edges } = action;